+++ src/Symfony/Component/HttpKernel/Tests/CacheWarmer/CacheWarmerTest.php
@@ -54,9 +54,14 @@ class TestCacheWarmer extends CacheWarmer
         $this->file = $file;
     }
 
+    /**
+     * @return string[]
+     */
     public function warmUp(string $cacheDir)
     {
         $this->writeCacheFile($this->file, 'content');
+
+        return [];
     }
 
     public function isOptional(): bool