+++ src/Symfony/Component/VarExporter/VarExporter.php
@@ -34,12 +34,13 @@ final class VarExporter
      *
      * @param mixed $value          The value to export
      * @param bool  &$isStaticValue Set to true after execution if the provided value is static, false otherwise
+     * @param bool  &$classes       Classes found in the value are added to this list as both keys and values
      *
      * @return string The value exported as PHP code
      *
      * @throws ExceptionInterface When the provided value cannot be serialized
      */
-    public static function export($value, bool &$isStaticValue = null): string
+    public static function export($value, bool &$isStaticValue = null, array &$foundClasses = []): string
     {
         $isStaticValue = true;
 
@@ -71,7 +72,9 @@ final class VarExporter
         $values = [];
         $states = [];
         foreach ($objectsPool as $i => $v) {
-            list(, $classes[], $values[], $wakeup) = $objectsPool[$v];
+            [, $class, $values[], $wakeup] = $objectsPool[$v];
+            $foundClasses[$class] = $classes[] = $class;
+
             if (0 < $wakeup) {
                 $states[$wakeup] = $i;
             } elseif (0 > $wakeup) {